pyproject.toml
[project]
name = "secretary"
version = "0.1.0"
description = "AI Family Secretary - 텔레그램/슬랙 기반 가족 비서"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-telegram-bot[all]>=22.0",
    "slack-bolt[async]>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10,<4",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "anyio>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/secretary"]

[tool.ruff]
target-version = "py312"
line-length = 100
